pyproject.toml
[project]
name = "telegramemojis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "av>=16.0.1",
    "Pillow",
    "python-magic",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.pytest.ini_options]
pythonpath = "src"
